+++ src/SEEK.AdPostingApi.Client/Models/StandoutAdvertisement.cs
@@ -2,6 +2,8 @@
 {
     public class StandoutAdvertisement
     {
+        public string BrandingId { get; set; }
+
         public int? LogoId { get; set; }
 
         public string[] Bullets { get; set; }